requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
asyncpg
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
